requirements.txt
# Core data science libraries
pandas>=2.0.0
numpy>=1.24.0

# Machine learning and preprocessing
scikit-learn>=1.3.0

# Web framework for UI
streamlit>=1.28.0

# HTTP requests for poster service
requests>=2.31.0

# Visualization (optional - fallback to streamlit built-in)
matplotlib>=3.7.0
seaborn>=0.12.0

# Dataset downloading
opendatasets>=0.1.22

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.20.0

# Additional useful libraries for data science
scipy>=1.10.0

# Optional: For environment variables
python-dotenv>=1.0.0

# Web UI and HTTP
streamlit>=1.36.0
requests>=2.32.0
python-dotenv>=1.0.1